Update SmoothScrollDelegate from scroll_bar.py
Default behaviour of scrollable widgets in PyQt and Qt is to propagate the scrolling event to parent widgets whenever their scrolling area is at its end. All of the widgets that used SmoothScrollDelegate would cause an issue when used inside another QScrollArea. By checking if the scrollbar is at its end, we can avoid this issue.